check if tee pid exists before killing
git-svn-id: https://tor.ern.nps.edu/svn/proj/labtainer/trunk@70 ed162e55-3153-4343-a18d-e086fa355ed4

+++ scripts/MyStudentDocker/bin/capinout.sh
@@ -81,7 +81,10 @@ else
 fi
 
 TEE_PID=$(ps | grep [t]ee | awk '{print $1}')
-kill $TEE_PID
+if [ ! -z "$TEE_PID" ]; then
+    kill $TEE_PID
+fi
+
 
 #exit ${PIPESTATUS[1]}
